Peer Roles working state
