--- frontend/src/contexts/peer_roles/components/PeerRoleTemplatesTable.tsx
@@ -1,51 +1,0 @@
-import { Stack, Table, Text } from "@mantine/core";
-import type { PeerRoleTemplate } from "../../../../api/Api";
-import { Anchor } from "../../../../components";
-
-interface PeerRoleTemplatesTableProps {
-  eventTemplates: PeerRoleTemplate[];
-}
-
-function PeerRoleTemplateTableRow({
-  template,
-}: {
-  template: PeerRoleTemplate;
-}) {
-  if (!template) return null;
-
-  return (
-    <Table.Tr key={template.id}>
-      <Table.Td>{template.id}</Table.Td>
-      <Table.Td>
-        <Stack gap={0}>
-          <Text>{template.name}</Text>
-          <Text c="dimmed">{template.summary}</Text>
-        </Stack>
-      </Table.Td>
-      <Table.Td>
-        <Anchor href={`${template.id}/edit`}>Edit</Anchor>
-      </Table.Td>
-    </Table.Tr>
-  );
-}
-
-export default function PeerRoleTemplatesTable({
-  peerroleTemplates,
-}: PeerRoleTemplatesTableProps) {
-  return (
-    <Table>
-      <Table.Thead>
-        <Table.Tr>
-          <Table.Th>#</Table.Th>
-          <Table.Th>Name</Table.Th>
-          <Table.Th>Actions</Table.Th>
-        </Table.Tr>
-      </Table.Thead>
-      <Table.Tbody>
-        {peerroleTemplates.map((template) => (
-          <PeerRoleTemplateTableRow key={template.id} template={template} />
-        ))}
-      </Table.Tbody>
-    </Table>
-  );
-}
